general_requirements.txt
# Python version: 3.12.4

streamlit-antd-components==0.3.2
streamlit==1.36.0
python-dotenv==1.0.0
pymongo==4.6.0
openai==1.2.4
streamlit-authenticator==0.2.3
st_clickable_images==0.0.3
python-dotenv==1.0.0
authlib==1.3.0
matplotlib==3.9.0
certifi==2023.11.17
flask==3.0.2
gunicorn==21.2.0
-e ./learnloop